chore: abstract chat & messages
